Put 'thread_ts' only when it's provided.
After #38, 'robot.say(...)' requires us to pass 'thread_ts' directly or
pass the original message with the 'thread_ts' attribute.
This change loosens the requirements and allows us to do
'robot.say(channel: "CHANNEL_ID", text: "Text")'
to post a simple message to a channel.
The behavior introduced in the PR #38 should not be changed by this
change.